requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
pydantic==2.10.5
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
python-dotenv==1.0.0
requests==2.31.0
